+++ mex/common/identity/base.py
@@ -2,7 +2,7 @@ from abc import abstractmethod
 
 from mex.common.connector import BaseConnector
 from mex.common.identity.models import Identity
-from mex.common.types import Identifier, PrimarySourceID
+from mex.common.types import AnyMergedIdentifier, MergedPrimarySourceIdentifier
 
 
 class BaseProvider(BaseConnector):
@@ -11,7 +11,7 @@ class BaseProvider(BaseConnector):
     @abstractmethod
     def assign(
         self,
-        had_primary_source: PrimarySourceID,
+        had_primary_source: MergedPrimarySourceIdentifier,
         identifier_in_primary_source: str,
     ) -> Identity:  # pragma: no cover
         """Find an Identity in a database or assign a new one."""
@@ -21,9 +21,9 @@ class BaseProvider(BaseConnector):
     def fetch(
         self,
         *,
-        had_primary_source: Identifier | None = None,
+        had_primary_source: MergedPrimarySourceIdentifier | None = None,
         identifier_in_primary_source: str | None = None,
-        stable_target_id: Identifier | None = None,
+        stable_target_id: AnyMergedIdentifier | None = None,
     ) -> list[Identity]:  # pragma: no cover
         """Find Identity instances matching the given filters."""
         ...