+++ mex/common/models/distribution.py
@@ -6,14 +6,15 @@ from mex.common.models.base import BaseModel
 from mex.common.models.extracted_data import ExtractedData
 from mex.common.models.merged_item import MergedItem
 from mex.common.types import (
-    AccessPlatformID,
     AccessRestriction,
-    DistributionID,
+    ExtractedDistributionIdentifier,
     License,
     Link,
+    MergedAccessPlatformIdentifier,
+    MergedDistributionIdentifier,
+    MergedOrganizationIdentifier,
+    MergedPersonIdentifier,
     MIMEType,
-    OrganizationID,
-    PersonID,
     YearMonth,
     YearMonthDay,
     YearMonthDayTime,
@@ -23,17 +24,16 @@ from mex.common.types import (
 class BaseDistribution(BaseModel):
     """A specific representation of a dataset."""
 
-    stableTargetId: DistributionID
-    accessService: AccessPlatformID | None = None
+    accessService: MergedAccessPlatformIdentifier | None = None
     accessRestriction: Annotated[
         AccessRestriction,
         Field(examples=["https://mex.rki.de/item/access-restriction-1"]),
     ]
     accessURL: Link | None = None
-    author: list[PersonID] = []
-    contactPerson: list[PersonID] = []
-    dataCurator: list[PersonID] = []
-    dataManager: list[PersonID] = []
+    author: list[MergedPersonIdentifier] = []
+    contactPerson: list[MergedPersonIdentifier] = []
+    dataCurator: list[MergedPersonIdentifier] = []
+    dataManager: list[MergedPersonIdentifier] = []
     downloadURL: Link | None = None
     issued: YearMonthDayTime | YearMonthDay | YearMonth
     license: (
@@ -49,11 +49,11 @@ class BaseDistribution(BaseModel):
         | None
     ) = None
     modified: YearMonthDayTime | YearMonthDay | YearMonth | None = None
-    otherContributor: list[PersonID] = []
-    projectLeader: list[PersonID] = []
-    projectManager: list[PersonID] = []
-    publisher: Annotated[list[OrganizationID], Field(min_length=1)]
-    researcher: list[PersonID] = []
+    otherContributor: list[MergedPersonIdentifier] = []
+    projectLeader: list[MergedPersonIdentifier] = []
+    projectManager: list[MergedPersonIdentifier] = []
+    publisher: Annotated[list[MergedOrganizationIdentifier], Field(min_length=1)]
+    researcher: list[MergedPersonIdentifier] = []
     title: Annotated[
         str,
         Field(
@@ -66,14 +66,17 @@ class BaseDistribution(BaseModel):
 class ExtractedDistribution(BaseDistribution, ExtractedData):
     """An automatically extracted metadata set describing a distribution."""
 
-    entityType: Literal["ExtractedDistribution"] = Field(
-        "ExtractedDistribution", alias="$type", frozen=True
-    )
+    entityType: Annotated[
+        Literal["ExtractedDistribution"], Field(alias="$type", frozen=True)
+    ] = "ExtractedDistribution"
+    identifier: Annotated[ExtractedDistributionIdentifier, Field(frozen=True)]
+    stableTargetId: MergedDistributionIdentifier
 
 
 class MergedDistribution(BaseDistribution, MergedItem):
     """The result of merging all extracted data and rules for a distribution."""
 
-    entityType: Literal["MergedDistribution"] = Field(
-        "MergedDistribution", alias="$type", frozen=True
-    )
+    entityType: Annotated[
+        Literal["MergedDistribution"], Field(alias="$type", frozen=True)
+    ] = "MergedDistribution"
+    identifier: Annotated[MergedDistributionIdentifier, Field(frozen=True)]