+++ mex/common/models/person.py
@@ -5,14 +5,19 @@ from pydantic import Field
 from mex.common.models.base import BaseModel
 from mex.common.models.extracted_data import ExtractedData
 from mex.common.models.merged_item import MergedItem
-from mex.common.types import Email, OrganizationalUnitID, OrganizationID, PersonID
+from mex.common.types import (
+    Email,
+    ExtractedPersonIdentifier,
+    MergedOrganizationalUnitIdentifier,
+    MergedOrganizationIdentifier,
+    MergedPersonIdentifier,
+)
 
 
 class BasePerson(BaseModel):
     """A person related to a source and/or resource, i.e. a project leader."""
 
-    stableTargetId: PersonID
-    affiliation: list[OrganizationID] = []
+    affiliation: list[MergedOrganizationIdentifier] = []
     email: list[Email] = []
     familyName: list[
         Annotated[
@@ -48,7 +53,7 @@ class BasePerson(BaseModel):
             ),
         ]
     ] = []
-    memberOf: list[OrganizationalUnitID] = []
+    memberOf: list[MergedOrganizationalUnitIdentifier] = []
     orcidId: list[
         Annotated[
             str,
@@ -64,14 +69,17 @@ class BasePerson(BaseModel):
 class ExtractedPerson(BasePerson, ExtractedData):
     """An automatically extracted metadata set describing a person."""
 
-    entityType: Literal["ExtractedPerson"] = Field(
-        "ExtractedPerson", alias="$type", frozen=True
-    )
+    entityType: Annotated[
+        Literal["ExtractedPerson"], Field(alias="$type", frozen=True)
+    ] = "ExtractedPerson"
+    identifier: Annotated[ExtractedPersonIdentifier, Field(frozen=True)]
+    stableTargetId: MergedPersonIdentifier
 
 
 class MergedPerson(BasePerson, MergedItem):
     """The result of merging all extracted data and rules for a person."""
 
-    entityType: Literal["MergedPerson"] = Field(
-        "MergedPerson", alias="$type", frozen=True
-    )
+    entityType: Annotated[
+        Literal["MergedPerson"], Field(alias="$type", frozen=True)
+    ] = "MergedPerson"
+    identifier: Annotated[MergedPersonIdentifier, Field(frozen=True)]